requirements.txt
flask==2.3.3
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
python-docx==1.2.0
werkzeug==2.3.7
